refactor(modularization): Phase 5 step 5-S4c — bot_service + QuotaOps DI (task #26)
Move ``bot_service`` into its canonical conversation-domain home and
swap its hard-import of ``aperag.service.quota_service`` for the
conversation-owned ``QuotaOps`` DI slot. Canonical msg=4a93c97e Q1
ruling C keeps ``aperag/service/quota_service.py`` as the legacy
provider through Phase 6 cleanup; ``aperag/app.py`` now wires the
legacy singleton into the new DI slot at module-scope import time
(same pattern as the Phase 3 Step 5b2c knowledge_base wire-ups).

Import rewrites — zero G1 legacy aggregate hits:

* ``from aperag.db import models as db_models`` +
  ``from aperag.db.models import Bot, BotStatus (, utc_now)`` (three
  inline imports inside the async closures) → single top-level
  ``from aperag.domains.conversation.db.models import Bot as BotRow,
  BotStatus, BotType`` + ``from aperag.utils.utils import utc_now``.
  The ORM row is aliased ``BotRow`` so the module does not shadow the
  Pydantic ``Bot`` response schema.
* ``from aperag.schema import view_models`` +
  ``from aperag.schema.view_models import Bot, BotList`` → explicit
  ``from aperag.domains.conversation.schemas import Bot, BotConfig,
  BotCreate, BotList, BotUpdate``.
* ``from aperag.service.quota_service import quota_service`` (G1
  banned) → ``from aperag.domains.conversation.ports import QuotaOps``
  + a module-level ``_quota_ops: QuotaOps | None`` with
  ``set_quota_ops`` / ``_get_quota_ops`` mirroring the Phase 3 Step
  5b2c pattern. ``_get_quota_ops`` raises ``RuntimeError`` when
  unwired instead of letting a ``None.check_and_consume_quota(...)``
  ``AttributeError`` leak out of the first request.

``aperag/service/bot_service.py`` collapses to a 36-line re-export
shim exposing ``BotService``, ``bot_service``, and ``set_quota_ops``
— the last one is what a caller outside ``aperag/app.py`` (for
example, the Phase 4 ``UserManager.on_after_register`` hook after it
rebases main) would use to swap the adapter without editing the
domain module.

``aperag/app.py`` gains:

* a new ``_conv_set_quota_ops`` import pulled through the legacy shim
  (explicit alias to keep the KB and conversation setters distinct
  at the wire-up site);
* a single ``_conv_set_quota_ops(_legacy_quota_service)`` call right
  after the four Phase 3 KB setters. The legacy ``quota_service``
  instance structurally satisfies both the KB and the conversation
  ``QuotaOps`` Protocols, so no new adapter class is needed — a
  follow-up ``QuotaInitOps`` adapter will appear in Phase 4 identity
  work (Bryce's branch), not here.

Verify:
* ``import aperag.app`` then
  ``aperag.domains.conversation.service.bot_service._quota_ops``
  is non-None (G17-style runtime smoke check passes today so the
  later 5-S8 CRITICAL_WIRINGS registry lands with zero surprises).
* ``from aperag.service.bot_service import BotService, bot_service,
  set_quota_ops`` OK via shim; ``BotService is
  aperag.domains.conversation.service.bot_service.BotService``.
* ``pytest tests/unit_test/`` 670 passed / 29 skipped / 1 deselected
  (performance flake pre-existing).
* ``uvx ruff check`` + ``uvx ruff format --check`` clean.
* G1 on ``aperag/domains/conversation/service/bot_service.py``: zero
  imports from the three legacy aggregate modules.
